sniffed_flooders_spoofers mac_arp_spoofer:
spoofed_arp() function sends a recived packet to target machine
using ip adress of machine we whis to impersonate leaving the
mac address argument blank. this then sends the packet to target
which uses the mac of our machine.

restore() function restors all mac address back to there original
state. The function is call when key is pressed.

previous error:
RuntimeError: Sniffing and sending packets is not available at layer 2: winpcap is not installed.
You may use conf.L3socket orconf.L3socket6 to access layer 3
I believe this was due to trying to print packet to console. I still am not
sure on how to do this or if it is even possible. It may be that packets need to be